requirements.txt
pyspark~=3.5.4
requests~=2.32.3
delta-spark
apache-airflow-providers-apache-spark